document-automation-python/requirements.txt
pandas
python-docx
docx2pdf
openpyxl
smtplib
email